starter/requirements.txt
pika
numpy
pandas
pyodbc
pyyaml
schedule



